Refactor code and merged with master to keep up-to-date
Moved RunBash to resources folder
TODO:
Integrate Buildjob with Runbash
See #19 for information
